Detect the cross-tissue eQTLs

Detect the cross-tissue eQTLs

Usage

select.ct.eQTL(
  x,
  nne = -1,
  gam = 1,
  nF0 = 5,
  verbose = T,
  allow_empty = T,
  ncores = 1L
)

Arguments

x

TWAS data, which is a list of three elements:

E

list of cell types, each element is a matrix of size n_sample x n_gene

dat

a list with three named elements.

bed

a matrix of size n_sample x n_snp

bim

a data frame with n_snp rows, each row is the info for the snp

pos

a list of genes, each element is a vector containing the SNP index that locate in the gene, and it corresponds to each row of E.info. It can be derived from E.info and dat.

nne

number of genes to be considered. Default is -1, which means to use all genes.

gam

the tuning parameter \(\gamma\) in the EBIC criterion. Default is 1.

nF0

the smallest number of main effects. Default is 5.

verbose

whether to print running message. Default is TRUE.

allow_empty

whether to allow empty set. Default is TRUE.

ncores

if using parallel computing, the number of cores to be used. Default is 1.

Value

a list of size n_gene, each element is the shared eQTL for the gene.