Select tissue-specific via SODA

Select tissue-specific via SODA

Usage

soda_model_imp_single(j, fix, E, dat, pos, gam = 1, scale = F, nF0 = 5)

Arguments

j

the index of gene

fix

a list of set containing the fixed effect for the gene

E

here E is just for specific cell type, it is a matrix instead of a list

dat

a list with three named elements.

bed

a matrix of size n_sample x n_snp

bim

a data frame with n_snp rows, each row is the info for the snp

pos

a list of genes, each element is a vector containing the SNP index that locate in the gene, and it corresponds to each row of E.info. It can be derived from E.info and dat.

gam

tuning parameter \(\gamma\) in EBIC criteria. Default is 1.

scale

whether to scale the expression and the genetic data when regressing the expression on the genetic data. Default is FALSE.

nF0

minimum size of the set